Fixes divide by zero error in getbottomrow.php
Since this premium thing is only in TF2 and I dont use the plugin with TF2 it always caused a PHP division by zero error. I am suggesting this fix -- it's working for me.

I didn't port this fix to the other things as they can't be zero and work in all sdk2013+ games.